% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpp}
\alias{gpp}
\title{GPP Sample Data}
\format{
\subsection{\code{gpp}}{

A tibble with 750 rows and 19 columns:
\describe{
\item{country}{Country name (Atlantis, Narnia, or Neverland)}
\item{year}{Survey year (2017, 2019, or 2022)}
\item{gend}{Sex of respondent (1 = Male, 2 = Female)}
\item{age}{Age of respondent}
\item{q1a, q1b, q1c, q1d, q1e, q1f}{Trust in institutions A to F
(1 = A lot, 2 = Some, 3 = A little, 4 = No trust,
99 = Don't know/No answer)}
\item{q49a, q49b_G1, q49b_G2, q49c_G1, q49c_G2, q49d_G1, q49d_G2,
q49e_G1, q49e_G2}{Confidence in different attributes of the criminal
justice system (1 = Very confident, 2 = Fairly confident,
3 = Not very confident, 4 = Not at all confident,
99 = Don't know/No answer). Items with the \verb{_G1} and \verb{_G2} suffixes
are question variants asked to different subsamples, and most 2017
respondents were not asked this battery, so these columns contain
many \code{NA} values.}
}
}
}
\source{
\url{https://worldjusticeproject.org/}
}
\usage{
gpp
}
\description{
A subset of data from the original General Population Poll (GPP).
Countries have been renamed into three fictional countries (Atlantis,
Narnia, Neverland) observed in three survey years.
}
\details{
Survey answers are stored as labelled vectors (\code{haven_labelled}, from
the haven package). Convert them to plain numbers before computing
indicators, e.g. \code{as.double(unclass(q1a))}.
}
\keyword{datasets}
