% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotsChart.R
\name{wjp_dots}
\alias{wjp_dots}
\title{Plot a Dots Chart following WJP style guidelines}
\usage{
wjp_dots(
  data,
  target,
  grouping,
  colors,
  cvec = NULL,
  order = NULL,
  diffOpac = FALSE,
  opacities = NULL,
  diffShp = FALSE,
  shapes = NA,
  draw_ci = FALSE,
  sd = NULL,
  sample_size = NULL,
  bgcolor = "#ffffff",
  ptheme = WJP_theme()
)
}
\arguments{
\item{data}{Data frame containing the data to plot}

\item{target}{String. Column name of the variable that will supply the values to plot.}

\item{grouping}{String. Column name of the variable that supplies the Y-Axis labels to show in the plot.}

\item{colors}{String. Column name of the variable that supplies the grouping values. The plot will show a different color per group.}

\item{cvec}{Named vector with the colors to apply to the dots. Default is NULL.}

\item{order}{String. Column name of the variable that contains the desired order for the labels.}

\item{diffOpac}{Boolean. If TRUE, the plot will expect different levels of opacities for the dots. Default is FALSE.}

\item{opacities}{Named vector with the opacity levels to apply to the dots. Default is NULL.}

\item{diffShp}{Boolean. If TRUE, the plot will expect different shapes for the dots. Default is FALSE.}

\item{shapes}{Named vector with shapes to be displayed. Default is NULL.}

\item{draw_ci}{Boolean. If TRUE, draws a normal-approximation confidence
interval using \code{sd} and \code{sample_size}.}

\item{sd}{String. Column name of the variable that supplies the standard error for drawing confidence intervals.}

\item{sample_size}{String. Column name of the variable that supplies the number of observations for drawing confidence intervals.}

\item{bgcolor}{String. Hex code for the "white" background in the strips.}

\item{ptheme}{ggplot theme function to apply to the plot. By default, function applies WJP_theme().}
}
\value{
A ggplot object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{wjp_dots()} takes a data frame with a specific data structure (usually long shaped) and returns a ggplot
object with a dots chart following WJP style guidelines.
}
\examples{
library(dplyr)
library(tidyr)
library(haven)
library(ggplot2)

# Always load the WJP fonts if not passing a custom theme to function
wjp_fonts()

# Preparing data
gpp_data <- WJPr::gpp

# Preparing data
data4dots <- gpp_data \%>\%
  select(country, q1a, q1b, q1c, q1d) \%>\%
  mutate(
    across(!country, \(x) as.double(unclass(x))),
    across(
      !country,
      \(x) case_when(
        x <= 2 ~ 1,
        x <= 4 ~ 0
      )
    )
  ) \%>\%
  group_by(country) \%>\%
  summarise(
    across(
      everything(),
      \(x) mean(x, na.rm = TRUE)*100
    ),
    .groups = "keep"
  ) \%>\%
  pivot_longer(
    !country,
    names_to  = "variable",
    values_to = "percentage" 
  ) \%>\%
  mutate(
    institution = case_when(
      variable == "q1a" ~ "Institution A",
      variable == "q1b" ~ "Institution B",
      variable == "q1c" ~ "Institution C",
      variable == "q1d" ~ "Institution D",
    )
  )

# Plotting chart
wjp_dots(
  data4dots,             
  target      = "percentage",
  grouping    = "institution",  
  colors      = "country",  
  cvec        = c("Atlantis"  = "#482d8b",
                  "Narnia"    = "#2894aa",
                  "Neverland" = "#f26b21")
)

}
