% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lollipopChart.R
\name{wjp_lollipops}
\alias{wjp_lollipops}
\title{Plot a Lollipop Chart following WJP style guidelines}
\usage{
wjp_lollipops(
  data,
  target,
  grouping,
  order = NULL,
  line_size = 3,
  point_size = 4,
  line_color = "#d9dde3",
  point_color = "#482d8b"
)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{target}{A string specifying the column name of the variable that contains the numeric values to be plotted.}

\item{grouping}{A string specifying the column name of the variable that contains the categories for the Y-axis labels.}

\item{order}{A string specifying the column name of the variable that contains the custom order for displaying categories. Default is NULL (uses row order).}

\item{line_size}{A numeric value specifying the thickness of the lines. Default is 3.}

\item{point_size}{A numeric value specifying the size of the points. Default is 4.}

\item{line_color}{A string specifying the hex color code for the lines. Default is "#d9dde3".}

\item{point_color}{A string specifying the hex color code for the points. Default is "#482d8b".}
}
\value{
A ggplot object representing the lollipop chart.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{wjp_lollipops()} takes a data frame with a specific data structure (usually long shaped) and returns a ggplot
object with a lollipop chart following WJP style guidelines.
}
\examples{
library(dplyr)
library(ggplot2)

# Always load the WJP fonts (optional)
wjp_fonts()

# Preparing data
gpp_data <- WJPr::gpp

data4lollipops <- gpp_data \%>\%
  filter(year == 2022, country == "Atlantis") \%>\%
  select(q1a, q1b, q1c, q1d) \%>\%
  mutate(
    across(everything(), \(x) as.double(unclass(x))),
    across(
      everything(),
      \(x) case_when(
        x <= 2 ~ 1,
        x <= 4 ~ 0
      )
    )
  ) \%>\%
  summarise(
    across(
      everything(),
      \(x) mean(x, na.rm = TRUE) * 100
    )
  ) \%>\%
  tidyr::pivot_longer(
    everything(),
    names_to  = "variable",
    values_to = "percentage"
  ) \%>\%
  mutate(
    institution = case_when(
      variable == "q1a" ~ "Institution A",
      variable == "q1b" ~ "Institution B",
      variable == "q1c" ~ "Institution C",
      variable == "q1d" ~ "Institution D"
    )
  )

# Plotting chart
wjp_lollipops(
  data4lollipops,
  target      = "percentage",
  grouping    = "institution",
  line_color  = "#d9dde3",
  point_color = "#482d8b"
)

}
