% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WJPr-package.R
\docType{package}
\name{WJPr-package}
\alias{WJPr}
\alias{WJPr-package}
\title{WJPr: World Justice Project Data and Visualization Tools}
\description{
The WJPr package contains the most important data and visualization tools used by the Data Analytics Unit at The World Justice Project. It provides a wide range of visualization functions for replicating the charts from the Country Reports following WJP style guidelines, a statistical helper for difference-in-means tests, and the Rule of Law Index scores for all factors and subfactors as complementary data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/worldjusticeproject/WJPr}
  \item \url{https://worldjusticeproject.github.io/WJPr/}
  \item Report bugs at \url{https://github.com/worldjusticeproject/WJPr/issues}
}

}
\author{
\strong{Maintainer}: World Justice Project Data Analytics Unit \email{data-analytics@worldjusticeproject.org} [copyright holder]

Authors:
\itemize{
  \item World Justice Project Data Analytics Unit \email{data-analytics@worldjusticeproject.org} [copyright holder]
  \item Santiago Pardo \email{spardo@worldjusticeproject.org} (github: aspardog)
}

}
\keyword{internal}
