% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpp}
\alias{gpp}
\title{GPP sample data}
\format{
\subsection{\code{gpp}}{

A data frame with 750 rows and 19 columns:
\describe{
\item{country}{Country name}
\item{year}{Year}
\item{gend}{Gender}
\item{age}{Age}
\item{q1x}{Trust in X institution}
\item{q49x}{Justice System evaluation}
...
}
}
}
\source{
\url{https://worldjusticeproject.org/}
}
\usage{
gpp
}
\description{
A subset of data from the original General Population Poll Data.
Countries and years have been renamed and group into three fictional
countries and years.
}
\keyword{datasets}
