% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{roli}
\alias{roli}
\title{Rule of Law Index Historical Data}
\format{
\subsection{\code{roli}}{

A data frame with 1,199 rows and 57 columns:
\describe{
\item{country}{Country name}
\item{year}{Year}
\item{roli}{Rule of Law Index Score}
\item{f1, f2, f3, f4, f5, f6, f7, f8}{Factor scores}
\item{sfxx}{Subfactor scores}
...
}
}
}
\source{
\url{https://worldjusticeproject.org/}
}
\usage{
roli
}
\description{
Index scores, at country-year level, for all factors and subfactors
from 2012 to 2023.
}
\keyword{datasets}
