% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divBars.R
\name{wjp_divbars}
\alias{wjp_divbars}
\title{Plot a Diverging Horizontal Bar Chart following WJP style guidelines}
\usage{
wjp_divbars(
  data,
  target,
  grouping,
  diverging,
  negative = NULL,
  cvec = NULL,
  labels = NULL,
  custom_order = F,
  order = NULL,
  ptheme = WJP_theme()
)
}
\arguments{
\item{data}{Data frame containing the data to plot}

\item{target}{String. Column name of the variable that will supply the values to plot.}

\item{grouping}{String. Column name of the variable that supplies the grouping values (Y-Axis Labels).}

\item{diverging}{String. Column name of the variable that supplies the diverging values.}

\item{negative}{String. Value that indicates that the bar should be in the negative quadrant.}

\item{cvec}{Named vector with the colors to apply to each bar segment. Default is NULL.}

\item{labels}{String. Column name of the variable that supplies the labels to show in the plot. Default is NULL.}

\item{custom_order}{Boolean. If TRUE, the plot will expect a custom order of the graph labels. Default is FALSE.}

\item{order}{String. Vector that contains the custom order for the y-axis labels. Default is NULL.}

\item{ptheme}{ggplot theme function to apply to the plot. By default, function applies WJP_theme()}
}
\value{
A ggplot object
}
\description{
\code{wjp_divbars()} takes a data frame with a specific data structure (usually long shaped) and returns a ggplot
object with a diverging horizontal bar chart following WJP style guidelines.
}
\examples{

\dontrun{
# Always load the WJP fonts if not passing a custom theme to function
wjp_fonts()

# Preparing data
data2plot <- gpp \%>\% 
  select(country, q49a)\%>\%
  mutate(div = case_when(
                   q49a == 1  ~ "Very confident",
                   q49a == 2  ~ "Fairly confident",
                   q49a == 3  ~ "Not very confident",
                   q49a == 4  ~ "Not at all confident",
                   q49a == 99 ~ NA_character_
               )
    )\%>\% drop_na()\%>\%
    group_by(country, div)\%>\%
count()\%>\%
group_by(country)\%>\%
transmute(div, value2plot = n/sum(n)*100)

data2plot$Label<- paste0(round(data2plot$value2plot,1), "\%")

colors4plot<- c("lightpink1", "palevioletred1", "violet", "mediumpurple1")
names(colors4plot)<- c("Very confident", "Fairly confident", "Not very confident", "Not at all confident")

data2plot$value2plot<- ifelse(data2plot$div \%in\% c("Not very confident", "Not at all confident"), data2plot$value2plot*-1, data2plot$value2plot)
data2plot$div<- as.factor(data2plot$div)
data2plot$div<- factor(data2plot$div, levels = c("Not at all confident", "Not very confident", "Very confident", "Fairly confident"))

# Plotting chart
wjp_divbars(data = data2plot, target = "value2plot", grouping = "country", diverging = "div", cvec = colors4plot, labels = "Label")
}
}
