% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumbbellsChart.R
\name{wjp_dumbbells}
\alias{wjp_dumbbells}
\title{Plot a Dumbbell Chart following WJP style guidelines}
\usage{
wjp_dumbbells(
  data,
  target,
  rows,
  color,
  cgroups,
  labels = NULL,
  labpos = NULL,
  cvec = NULL,
  order = NULL,
  ptheme = WJP_theme()
)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{target}{A string specifying the variable in the data frame that contains the numeric values to be plotted.}

\item{rows}{A string specifying the variable in the data frame that contains the categories for the rows.}

\item{color}{A string specifying the variable in the data frame that indicates the groups for start and end points.}

\item{cgroups}{A vector of two strings specifying the groups to be compared in the dumbbell plot.}

\item{labels}{A string specifying the variable in the data frame that contains the text labels to display.}

\item{labpos}{A string specifying the variable in the data frame that contains the label positions.}

\item{cvec}{A vector of colors to apply to the points and lines. Default is NULL.}

\item{order}{A named vector specifying the order of the categories. Default is NULL.}

\item{ptheme}{A ggplot2 theme object to be applied to the plot. Default is WJP_theme().}
}
\value{
A ggplot object representing the dumbbell plot.
}
\description{
\code{wjp_dumbbells()} takes a data frame with a specific data structure (usually long shaped) and returns a ggplot
object with a dumbbell chart following WJP style guidelines.
}
\examples{
\dontrun{
data <- data.frame(
  group = c("A", "B", "C", "A", "B", "C"),
  target = c(30, 50, 20, 40, 60, 30),
  color = c("start", "start", "start", "end", "end", "end")
)
cgroups <- c("start", "end")
cvec <- c("#2a2a94", "#d1cfd1")

wjp_dumbbells(data, "target", "group", "color", cgroups, cvec)
}

}
