% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgebarsChart.R
\name{wjp_edgebars}
\alias{wjp_edgebars}
\title{Plot a Horizontal Edgebars Chart following WJP style guidelines}
\usage{
wjp_edgebars(
  data,
  y_value,
  x_var,
  label_var,
  color_var = NULL,
  bar_colors = NULL,
  x_lab_pos = NULL,
  y_lab_pos = 0,
  nudge_lab = 2.5,
  margin_top = 20,
  bar_width = 0.35,
  ptheme = WJP_theme()
)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{y_value}{A string specifying the variable in the data frame that contains the numeric values to be plotted as bars.}

\item{x_var}{A string specifying the variable in the data frame that contains the categories for the x-axis.}

\item{label_var}{A string specifying the variable in the data frame that contains the labels to be displayed near the bars.}

\item{color_var}{A string specifying the variable in the data frame that contains the the color groupings for the bars. Default is NULL.}

\item{bar_colors}{A named vector specifying the colors for the bars. Default is NULL.}

\item{x_lab_pos}{A string specifying the variable in the data frame that contains the order in which the bars will be displayed. Default is NULL.}

\item{y_lab_pos}{A numeric value specifying the y-axis position for displaying labels. Default is 0.}

\item{nudge_lab}{A numeric value specifying the padding for displaying labels in milimeters. Default is 2.5.}

\item{margin_top}{A numeric value specifying the top margin of the plot. Default is 20.}

\item{bar_width}{A numeric value specifying the width of the bars. For single bars the default value of 0.35 is recommended, for plots with two bars a value of 0.5 is more suitable.}

\item{ptheme}{A ggplot aesthetic theme to be applied to the chart. Default is the WJP_theme initilized with the package.}
}
\value{
A ggplot object representing the edge bars plot.
}
\description{
\code{wjp_edgebars()} takes a data frame with a specific data structure (usually long shaped) and returns a ggplot
object with an edgebar chart following WJP style guidelines.
}
\examples{
\dontrun{
data <- data.frame(
  y_value = c(30, 50, 20),
  x_var = c("Category A", "Category B", "Category C"),
  label_var = c("Label A", "Label B", "Label C")
)

wjp_edgebars(data, "y_value", "x_var", "label_var")
}


}
