% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaugeChart.R
\name{wjp_gauge}
\alias{wjp_gauge}
\title{Plot a Gauge Chart following WJP style guidelines}
\usage{
wjp_gauge(
  data,
  target,
  colors,
  cvec = NULL,
  factor_order = NULL,
  labels = NULL,
  crop = c(-10, 0, 0, -8),
  ptheme = WJP_theme()
)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{target}{A string specifying the variable in the data frame that contains the values to be plotted.}

\item{colors}{A string specifying the variable in the data frame that represents the color groupings for the segments.}

\item{cvec}{A vector of colors to apply to the segments.}

\item{factor_order}{A vector specifying the order in which the segments should be plotted.}

\item{labels}{A string specifying the variable in the data frame that contains the labels to be displayed. Default is NULL.}

\item{crop}{A numeric vector specifying the amount of space to crop from the Top, Right, Bottom, and Right margins, respectively. Default is c(-10,0,0,-8).}

\item{ptheme}{A ggplot2 theme object to be applied to the plot. Default is WJP_theme().}
}
\value{
A ggplot object representing the gauge chart.
}
\description{
This function creates a gauge chart using ggplot2 based on the provided data frame.
}
\examples{
\dontrun{
data <- data.frame(
  target = c(10, 20, 30, 40),
  colors = c("red", "blue", "green", "yellow"),
  labels = c("A", "B", "C", "D")
)
cvec <- c("red", "blue", "green", "yellow")

wjp_gauge(data, "target", "colors", cvec, labels = "labels")
}


}
