% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarChart.R
\name{wjp_radar}
\alias{wjp_radar}
\title{Plot a Radar Chart following WJP style guidelines}
\usage{
wjp_radar(
  data,
  axis_var,
  target_var,
  label_var,
  color_var,
  colors,
  order_var,
  maincat,
  source = "GPP"
)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{axis_var}{A string specifying the variable in the data frame that contains the groups for the axis.}

\item{target_var}{A string specifying the variable in the data frame that contains the values to be plotted.}

\item{label_var}{A string specifying the variable in the data frame that contains the labels to be displayed.}

\item{color_var}{A string specifying the variable in the data frame that contains the color groupings.}

\item{colors}{A vector of colors to apply to lines. The first color will be applied to percentages in labels.}

\item{order_var}{A string specifying the variable in the data frame that contains the display order of categories.}

\item{maincat}{A string indicating the category of labels to show in the radar.}

\item{source}{A string which can take two values (GPP or QRQ).}
}
\value{
A ggplot object representing the radar plot.
}
\description{
#' @description
\code{wjp_radar()} takes a data frame with a specific data structure (usually long shaped) and returns a ggplot
object with a radar chart following WJP style guidelines.
}
\examples{
\dontrun{
data <- data.frame(
  axis_var = rep(letters[1:5], each = 3),
  target_var = runif(15, 0, 1),
  label_var = rep(LETTERS[1:5], each = 3),
  order_var = rep(1:5, each = 3),
  year = rep(2020:2022, 5)
)
colors <- c("red", "blue", "green")
maincat <- 2022

wjp_radar(data, "axis_var", "target_var", "label_var", "order_var", colors, maincat)
}


}
