% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarChart.R
\name{wjp_radar}
\alias{wjp_radar}
\title{Plot a Radar Chart following WJP style guidelines}
\usage{
wjp_radar(
  data,
  axis_var,
  target_var,
  label_var,
  color_var,
  maincat,
  cvec = NULL,
  order_var = NULL,
  source = "GPP"
)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{axis_var}{A string specifying the variable in the data frame that contains the groups for the axis.}

\item{target_var}{A string specifying the variable in the data frame that contains the values to be plotted.}

\item{label_var}{A string specifying the variable in the data frame that contains the labels to be displayed.}

\item{color_var}{A string specifying the variable in the data frame that contains the color groupings.}

\item{maincat}{A string indicating the category of labels to show in the radar.}

\item{cvec}{A vector of colors to apply to lines. The first color will be applied to percentages in labels.}

\item{order_var}{A string specifying the variable in the data frame that contains the display order of categories. Default is NULL.}

\item{source}{A string which can take two values (GPP or QRQ).}
}
\value{
A ggplot object representing the radar plot.
}
\description{
#' @description
\code{wjp_radar()} takes a data frame with a specific data structure (usually long shaped) and returns a ggplot
object with a radar chart following WJP style guidelines.
}
\examples{
# Always load the WJP fonts (optional)
wjp_fonts()

# Preparing data
gpp_data <- WJPr::gpp

library(dplyr)
library(tidyr)
library(haven)

data4radar <- gpp_data \%>\%
select(gend, starts_with("q49")) \%>\%
  mutate(
    gender = case_when(
      gend == 1 ~ "Male",
      gend == 2 ~ "Female"
    ),
    across(
      starts_with("q49"),
      \(x) case_when(
        x <= 2  ~ 1,
        x <= 99 ~ 0
      )
    )
  ) \%>\%
  group_by(gender) \%>\%
  summarise(
    across(
      starts_with("q49"),
      \(x) mean(x, na.rm = T)*100
    )
  ) \%>\%
  pivot_longer(
    !gender,
    names_to  = "category",
    values_to = "percentage"
  ) \%>\%
  mutate(
    axis_label = category
  )

# Plotting chart
wjp_radar(
  data4radar,             
 axis_var    = "category",         
  target_var  = "percentage",       
  label_var   = "axis_label",        
  color_var   = "gender"
)

}
