% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarChart.R
\name{wjp_radar}
\alias{wjp_radar}
\title{Plot a Radar Chart following WJP style guidelines}
\usage{
wjp_radar(
  data,
  axis_var,
  target,
  labels,
  colors,
  maincat = NULL,
  cvec = NULL,
  order_var = NULL,
  source = "GPP"
)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{axis_var}{A string specifying the variable in the data frame that contains the groups for the axis.}

\item{target}{A string specifying the variable in the data frame that contains the values to be plotted.}

\item{labels}{A string specifying the variable in the data frame that contains the labels to be displayed.}

\item{colors}{A string specifying the variable in the data frame that contains the color groupings.}

\item{maincat}{A string indicating the column used to choose the axis labels.
If NULL, labels are taken from the first color group.}

\item{cvec}{A named vector of colors to apply to lines.}

\item{order_var}{A string specifying the variable in the data frame that contains the display order of categories. Default is NULL.}

\item{source}{A string which can take two values (GPP or QRQ).}
}
\value{
A ggplot object representing the radar plot.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{wjp_radar()} takes a data frame with a specific data structure (usually long shaped) and returns a ggplot
object with a radar chart following WJP style guidelines.
}
\examples{
library(dplyr)
library(tidyr)
library(haven)
library(ggplot2)
library(purrr)
library(ggtext)

# Always load the WJP fonts (optional)
wjp_fonts()

# Preparing data
gpp_data <- WJPr::gpp

data4radar <- gpp_data \%>\%
select(gend, starts_with("q49")) \%>\%
  mutate(
    gender = case_when(
      gend == 1 ~ "Male",
      gend == 2 ~ "Female"
    ),
    across(
      starts_with("q49"),
      \(x) case_when(
        x <= 2  ~ 1,
        x <= 99 ~ 0
      )
    )
  ) \%>\%
  group_by(gender) \%>\%
  summarise(
    across(
      starts_with("q49"),
      \(x) mean(x, na.rm = T)*100
    )
  ) \%>\%
  pivot_longer(
    !gender,
    names_to  = "category",
    values_to = "percentage"
  ) \%>\%
  mutate(
    axis_label = category
  )

# Plotting chart
wjp_radar(
  data4radar,             
  axis_var    = "category",         
  target      = "percentage",       
  labels      = "axis_label",        
  colors      = "gender"
)

}
