% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WJPr-package.R
\docType{package}
\name{WJPr-package}
\alias{WJPr}
\alias{WJPr-package}
\title{WJPr: World Justice Project data and visualizations tools}
\description{
The WJPr package contains the most important data and visualization tools used by Data Analytics Unit at The World Justice Project. Version 1.0.0 comes with a wide range of visualization functions for replicating the image charts from the Country Reports. Additionally, you can find the Rule of Law Index scores data for all factors and subfactors as complementary data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ctoruno/WJPr}
  \item \url{https://ctoruno.github.io/WJPr/}
  \item Report bugs at \url{https://github.com/ctoruno/WJPr/issues}
}

}
\author{
\strong{Maintainer}: Carlos A. Toruno Paniagua \email{ctoruno@worldjusticeproject.org}

}
\keyword{internal}
