% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WJPr-package.R
\docType{package}
\name{WJPr-package}
\alias{WJPr}
\alias{WJPr-package}
\title{WJPr: World Justice Project Data and Visualization Tools}
\description{
The WJPr package contains the most important data and visualization tools used by Data Analytics Unit at The World Justice Project. Version 1.0.0 comes with a wide range of visualization functions for replicating the image charts from the Country Reports. Additionally, you can find the Rule of Law Index scores data for all factors and subfactors as complementary data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/worldjusticeproject-org/WJPr}
  \item \url{https://worldjusticeproject-org.github.io/WJPr/}
  \item Report bugs at \url{https://github.com/worldjusticeproject-org/WJPr/issues}
}

}
\author{
\strong{Maintainer}: Santiago Pardo \email{spardo@worldjusticeproject.org} (github: aspardog) [copyright holder]

Authors:
\itemize{
  \item Santiago Pardo \email{spardo@worldjusticeproject.org} (github: aspardog) [copyright holder]
  \item World Justice Project Data Analytics Unit \email{data-analytics@worldjusticeproject.org} [copyright holder]
}

}
\keyword{internal}
